requirements.txt
Django==1.6.1
Markdown==2.3.1
South==0.8.4
Unidecode==0.04.14
amqp==1.4.3
anyjson==0.3.3
billiard==3.3.0.16
celery==3.1.9
django-facebook==5.3.1
django-filter==0.7
djangorestframework==2.3.12
gunicorn==18.0
ipython==1.1.0
kombu==3.0.12
pytz==2013.9
simplejson==3.3.2
wsgiref==0.1.2
